requirements.txt
fastapi
websockets
uvicorn
python-dotenv
supabase
streamlit
requests